tests(Form): add an example for disabled state in cypress
also remove example from Form
